Fix JSON parsing error and remove unused UI elements
Fix bbox coordinate parsing from backend Float to Double type.
Backend returns bbox coordinates as Floats but DTOs expected Int.
Changed DetectionDTO.bbox and Detection.bbox from List<Int> to List<Double>.

Remove Generate Report button from Reports screen header.
Report generation is now handled via New Analysis module.

Platforms: KMP (Desktop, Web, Android)

+++ composeApp/src/commonMain/kotlin/com/dentalvision/ai/data/remote/api/dto/AnalysisReportDTO.kt
@@ -35,7 +35,7 @@ data class DetectionDTO(
     @SerialName("class_name")
     val className: String,
     val confidence: Double,
-    val bbox: List<Int>, // [x1, y1, x2, y2]
+    val bbox: List<Double>, // [x1, y1, x2, y2]
     val probabilities: List<Double>
 )
 
@@ -91,7 +91,7 @@ data class Detection(
     val classId: Int,
     val className: String,
     val confidence: Double,
-    val bbox: List<Int>,
+    val bbox: List<Double>,
     val probabilities: List<Double>
 ) {
     fun getConfidencePercentage(): String = "${(confidence * 100).toInt()}%"
